Fixes #84: add support of access log cleaning shutdown.

+++ protected/commands/ClearAccessLogCommand.php
@@ -2,9 +2,14 @@
 
 class ClearAccessLogCommand extends CConsoleCommand {
 	public function run() {
+		$access_log_lifetime = Parameters::getModel()->access_log_lifetime_in_s;
+		if (intval($access_log_lifetime) === 0) {
+			return;
+		}
+
 		Access::model()->deleteAll(
 			'timestamp < SUBDATE(NOW(), INTERVAL '
-			. Parameters::getModel()->access_log_lifetime_in_s
+			. $access_log_lifetime
 			. ' SECOND)'
 		);
 	}
